+++ src/OpenTelemetry.Exporter.Zipkin/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ## Unreleased
 
+* **Breaking change**: Non-primitive tag values converted using
+  `Convert.ToString` will now format using `CultureInfo.InvariantCulture`.
+  ([#5700](https://github.com/open-telemetry/opentelemetry-dotnet/pull/5700))
+
+## 1.9.0
+
+Released 2024-Jun-14
+
+## 1.9.0-rc.1
+
+Released 2024-Jun-07
+
 ## 1.9.0-alpha.1
 
 Released 2024-May-20